requirements.txt
playwright
python-dotenv
openai
pyperclip
requests
Pillow